Rename to json_schema
